extract the sorting functions to another file
