+++ packages/visual-stack/src/components/Table/DataTable/index.js
@@ -11,43 +11,10 @@ import {
   Th,
 } from '../';
 import Pagination from '../../Pagination';
-import ArrowDownIcon from 'mdi-react/ArrowDownIcon';
-import ArrowUpIcon from 'mdi-react/ArrowUpIcon';
 import * as R from 'ramda';
+import { getNextData, getNextOrder, getSortingIcon } from './sortingHelper';
 import PropTypes from 'prop-types';
 import './DataTable.css';
-
-export const ASCENDING = 'ascending';
-export const DESCENDING = 'descending';
-
-const getSortingIcon = (sortingOption, currentLabel) => {
-  let icon = null;
-  if (sortingOption.order === ASCENDING) {
-    icon = <ArrowUpIcon />;
-  }
-  if (sortingOption.order === DESCENDING) {
-    icon = <ArrowDownIcon />;
-  }
-  if (sortingOption.label !== currentLabel) {
-    icon = <ArrowDownIcon className="vs-unsorted-row-icon" />;
-  }
-  return icon;
-};
-
-const getNextOrder = order => {
-  if (!order || order === ASCENDING) return DESCENDING;
-  if (order === DESCENDING) return ASCENDING;
-};
-
-const sort = order => index => R.sortWith([order(R.prop(index))]);
-const sortAscendingByIndex = sort(R.ascend);
-const sortDescendingByIndex = sort(R.descend);
-
-const getNextData = (index, currentOrder, data) => {
-  if (!currentOrder || currentOrder === ASCENDING)
-    return sortDescendingByIndex(index)(data);
-  if (currentOrder === DESCENDING) return sortAscendingByIndex(index)(data);
-};
 
 const generateHeader = ({ sortable, sortingOption, onSort, data }) => (
   { label: currentLabel, width },
